Centralize decoder, make tests more robust against timezones
